+++ CHANGELOG.md
@@ -37,6 +37,7 @@ The format is based on [Keep a Changelog].
 - Bones swung by unused PhysBones (which will be removed by AAO) are not merged `#850`
   - Note that To fix this problem, AnimatorParser is almost completely rewritten.
   - It's not expected to have behavior change, but if you found some, please report it.
+- Re-fix Nested Constraint can be broken with Trace and Optimize `#880`
 
 ### Security